refactor(e2e-native): split and move e2e utis methods
to break circular dep
fixes one missing await
