+++ suite-native/app/e2e/pageObjects/homeActions.ts
@@ -1,6 +1,6 @@
 import { expect as detoxExpect } from 'detox';
 
-import { waitForElementByIdToBeVisible } from '../utils';
+import { waitForElementByIdToBeVisible } from '../support/utils';
 
 const graphHeaderDiscreetTextElement = element(
     by.id('@screen/Home').withDescendant(by.id('discreet-text')),
