+++ suite-native/app/e2e/support/setup.ts
@@ -1,4 +1,3 @@
-import { expect as detoxExpect } from 'detox';
 import { resolveConfig } from 'detox/internals';
 
 import { LaunchArguments } from '@suite-native/config';
@@ -6,14 +5,16 @@ import { PreloadedState } from '@suite-native/state';
 import { MNEMONICS, MODELS, Model, TrezorUserEnvLink } from '@trezor/trezor-user-env-link';
 import { mergeDeepObject } from '@trezor/utils';
 
-import { onDeviceOnboarding } from './pageObjects/deviceOnboardingActions';
-import { onDevicePrompt } from './pageObjects/devicePromptActions';
+import { appIsFullyLoaded, platform } from './utils';
+import { onDeviceOnboarding } from '../pageObjects/deviceOnboardingActions';
+import { onDevicePrompt } from '../pageObjects/devicePromptActions';
 
-const platform = device.getPlatform();
-
-// There is inconsistency between platforms. Android needs to have 100% of an element visible to be able to interact with it.
-// On the other hand, if we are trying to scroll to 100% visibility on iOS, it causes scrolling more than height of the screen and it makes Detox crash.
-const SCROLL_VISIBILITY_THRESHOLD = platform === 'android' ? 100 : undefined;
+type PrepareTrezorEmulatorProps = {
+    seed?: string;
+    passphrase_protection?: boolean;
+    model?: Model;
+    version?: string;
+};
 
 const INITIAL_LAUNCH_ARGS: LaunchArguments = {
     // Do not synchronize communication with the trezor bridge and metro server running on localhost. Since the trezor
@@ -29,16 +30,6 @@ const INITIAL_LAUNCH_ARGS: LaunchArguments = {
 };
 
 const TREZOR_E2E_DEVICE_LABEL = 'Trezor T - Tester';
-
-export const wait = async (ms: number) => {
-    await new Promise(resolve => setTimeout(resolve, ms));
-};
-
-export const appIsFullyLoaded = async () => {
-    await waitFor(element(by.id('@screen/mainScrollView')))
-        .toBeVisible()
-        .withTimeout(35000);
-};
 
 const getExpoDeepLinkUrl = () => {
     const expoLauncherUrl = encodeURIComponent(
@@ -121,40 +112,11 @@ export const openApp = async ({
     }
 };
 
-export const scrollUntilVisible = async (
-    target: Detox.IndexableNativeElement,
-    scrollViewTestId: string = '@screen/mainScrollView',
-) => {
-    try {
-        // Try to confirm that the element is visible without scrolling.
-        await detoxExpect(target).toBeVisible(SCROLL_VISIBILITY_THRESHOLD);
-    } catch {
-        // If the element is not visible, then use the scroll to find it.
-        const scrollViewElement = element(by.id(scrollViewTestId));
-        await waitFor(scrollViewElement).toBeVisible().withTimeout(5000);
-
-        await waitFor(target)
-            .toBeVisible(SCROLL_VISIBILITY_THRESHOLD)
-            .whileElement(by.id(scrollViewTestId))
-            .scroll(300, 'down', 0.5, 0.5);
-
-        // wait for scroll animation to finish before performing next action
-        await wait(1000);
-    }
-};
-
 export function getModelFromEnv(): Model {
     const envValue = process.env.EMULATOR_MODEL as Model;
 
     return MODELS.includes(envValue) ? envValue : 'T3T1';
 }
-
-export type PrepareTrezorEmulatorProps = {
-    seed?: string;
-    passphrase_protection?: boolean;
-    model?: Model;
-    version?: string;
-};
 
 const getFwVersion = (model: Model, version: string | undefined) => {
     if (model === 'T3W1') {
@@ -206,16 +168,6 @@ export const prepareTrezorEmulator = async ({
     }
 };
 
-export const waitForElementByTextToBeVisible = (text: string, timeout = 30000) =>
-    waitFor(element(by.text(text)))
-        .toBeVisible()
-        .withTimeout(timeout);
-
-export const waitForElementByIdToBeVisible = (testId: string, timeout = 30000) =>
-    waitFor(element(by.id(testId)))
-        .toBeVisible()
-        .withTimeout(timeout);
-
 /**
  * Merges multiple preloaded state fragments into a single preloaded state and serializes the result.
  * Be mindful about the order of the fragments, as the later fragments will always override the earlier ones!
@@ -225,16 +177,4 @@ export const preparePreloadedReduxState = (...stateFragments: PreloadedState[]):
     const mergedState = mergeDeepObject(...definedFragments);
 
     return JSON.stringify(mergedState);
-};
-
-export const inputTextToElement = async (element: Detox.IndexableNativeElement, text: string) => {
-    // on Android it is very slow to type text symbol by symbol, for performance reasons `replaceText` is used instead.
-    if (platform === 'android') {
-        await element.replaceText(text);
-    } else {
-        // on iOS the replaceText do not trigger input events (focus, blur, etc.) so we need can not paste text there as for Android.
-        // the typeText method is way faster than for Android, so there is not performance drawback.
-        await element.tap();
-        await element.typeText(text);
-    }
 };
