+++ suite-native/app/e2e/tests/accountManagement.test.ts
@@ -10,7 +10,7 @@ import { onAccountDetailSettings } from '../pageObjects/accountDetailSettingsAct
 import { onHome } from '../pageObjects/homeActions';
 import { onMyAssets } from '../pageObjects/myAssetsActions';
 import { onTabBar } from '../pageObjects/tabBarActions';
-import { openApp, preparePreloadedReduxState } from '../utils';
+import { openApp, preparePreloadedReduxState } from '../support/setup';
 
 const preloadedState = preparePreloadedReduxState(
     onboardingCompletedState,
