+++ suite-native/app/e2e/tests/coinEnabling.test.ts
@@ -13,7 +13,7 @@ import {
     openApp,
     preparePreloadedReduxState,
     prepareTrezorEmulator,
-} from '../utils';
+} from '../support/setup';
 
 const preloadedState = preparePreloadedReduxState(
     onboardingCompletedState,
